pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Erick",email = "98369791+justthebasic@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "poetry-plugin-dotenv (>=2.9.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "together (>=1.5.5,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
